Initial submission
Perl modules (*.pm) and examples of perl scripts (*.pl) that use the
Poly3dReader module to parse Poly3D intput/output files.  Perl scripts
are all run from the command line, and will probably require the
FindBin module in the perl libraries.  See the comments in the
respective perl scripts, in the Poly3dReader.pm file, and in the README
file for instructions on usage.